refactor: student dashboard naming convention
